% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{color2IntV}
\alias{color2IntV}
\title{color2IntV}
\usage{
color2IntV(color)
}
\arguments{
\item{color}{A color specified either by name (e.g. "red") or as a RGB hexadecimal value (e.g. "#FF0000").}
}
\value{
A integer vector containing the RGB specification.
}
\description{
Convert a color, either as a name or as a RGB hexadecimal value to an integer vector containing
the RGB specification.
}
\seealso{
\code{\link[=intV2Color]{intV2Color()}}
}
