% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat2UnitNormRank.R
\name{complement}
\alias{complement}
\title{complement}
\usage{
complement(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
The complement of the x argument, equal to \eqn{1 - x}.
}
\description{
This function returns the complement of a numeric or integer vector or matrix. This may be suitable as the
\code{matrix_scaling_fun()} argument for \code{gsnPareNetGenericHierarchic()} when being used with such distance metrics
as the Jaccard Index or Szymkiewicz–Simpson Overlap Coefficients to transform them into something more approximating a
distance in behavior.
}
\details{
This function also sets matrix or vector attributes appropriately for negation of the input.
}
