% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_color.R
\name{contrasting_color}
\alias{contrasting_color}
\title{contrasting_color}
\usage{
contrasting_color(
  col,
  type = "complement",
  threshold = 127,
  low = "#000000",
  high = "#FFFFFF"
)
}
\arguments{
\item{col}{A character vector of colors.}

\item{type}{(optional) Type of contrasting color, i.e. the method of generating the contrasting color.
Valid values are \code{'complement'}, \code{'rotate'}, \code{'yellow'}, \code{'gray'}, \code{'binary'}
and \code{'blackyellow'}.}

\item{threshold}{(optional, used only for \code{type='binary'}) The "binary" method works by assessing
the mean value of the RGB channels. If the value is above a threshold, the low color is returned, if it
is below the threshold, the high color is returned.}

\item{low}{(optional, used only for \code{type='binary'}) Low color (see \code{threshold} argument).}

\item{high}{(optional, used only for \code{type='binary'}) High color (see \code{threshold} argument).}
}
\value{
A contrasting color.
}
\description{
Function picks colors to contrast with the colors given as the \code{'col'} argument.
}
