% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termSummary.R
\name{filterTerms}
\alias{filterTerms}
\title{filterTerms}
\usage{
filterTerms(terms, .stopwords = get_stopwords(), convert_underscores = TRUE)
}
\arguments{
\item{terms}{A vector of terms.}

\item{.stopwords}{A vector of stopwords (defaulting to the value returned by the private
function \code{get_stopwords()}.}

\item{convert_underscores}{A single logical value indicating whether or not underscores
should be converted to spaces. (default: TRUE)}
}
\value{
A character vector containing filtered (but not summarized) terms.
}
\description{
This function preprocesses a vector of terms, removing stopwords, converting
to lower case, and optionally converting underscores (as found in term names) to spaces.
}
\examples{
library( GSNA )
unfiltered_terms <- c("GOBP_COMPLEMENT_ACTIVATION", "REACTOME_COMPLEMENT_CASCADE",
         "GOMF_COMPLEMENT_BINDING", "WP_COMPLEMENT_ACTIVATION_CLASSICAL_PATHWAY",
         "GOBP_COMPLEMENT_ACTIVATION_CLASSICAL_PATHWAY", "BIOCARTA_LECTIN_PATHWAY",
         "GOMF_OPSONIN_BINDING", "GOCC_GOLGI_LUMEN" )
filterTerms( unfiltered_terms )

}
