% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsn_suggest_cutoff.R
\name{gsn_suggest_cutoff}
\alias{gsn_suggest_cutoff}
\title{gsn_suggest_cutoff}
\usage{
gsn_suggest_cutoff(
  object,
  distance = GSNA::gsn_default_distance(object),
  lower_is_closer = (function(x) {
     ifelse(is.null(x), TRUE, x)
 })(attr(x =
    object$distances[[distance]]$matrix, which = "lower_is_closer", exact = TRUE)),
  quantile = ifelse(lower_is_closer, yes = 0.25, no = 0.75),
  alpha = NULL
)
}
\arguments{
\item{object}{A GSNData object.}

\item{distance}{(optional) Specifies the distance metric to be used. (detault is the default distance.)}

\item{lower_is_closer}{(optional) Logical value specifying whether lower values are to be considered
closer, within the distance matrix. This should generally be TRUE for statistics like \emph{p}-values,
but FALSE for metrics such as Jaccard, OC, or Kappa, for which higher values are closer, i.e. more
overlap. (default is the lower_is_closer attribute of the distance matrix, or TRUE if that is not set.)}

\item{quantile}{(optional) The quantile to be used for specifying a distance cutoff. Defaults to 0.25
for lower_is_closer == TRUE and 0.75 for lower_is_closer == FALSE.}

\item{alpha}{(optional) Equivalent \emph{p}-value cutoff. This can be used when the distance matrix
distance_type attribute is "ln_pval". If the distance matrix is a ln-transformed \emph{p}-value and alpha is
less than the quantile based cutoff, a message indicating that the corresponding \emph{p}-value for the
specified quantile is not significant will be generated, and the quantile based cutoff will be overridden
and the specified \emph{p}-value based cutoff will be indicated instead. If not specified or set to \code{NULL},
the quantile-based cuttoff will be returned, even if it does not correspond to a significant \emph{p}-value.
In that case, a message will still indicate that it is not significant. For similarity metrics such as
Jaccard, Szymkiewicz-Simpson OC and Cohen's kappa, a corresponding \emph{p}-value cannot easily be calculated,
so this argument is not used.}
}
\value{
Invisibly returns a numerical vector indicating a suggested cutoff, based on the specified distance,
quantile values, and alpha (if appropriate).
}
\description{
A function that suggests a paring cutoff value for a GSNData object. This value can
optionallybe used as the \code{cutoff} argument for \code{\link{gsnPareNetGenericHierarchic}()} or
\code{\link{gsnPareNetGenericToNearestNNeighbors}()}.
}
\details{
In addition to the returned value, this function prints messages with various information,
including the distance matrix being used, whether lower values represent closer gene sets
(lower_is_closer, as specified in the function call and in the distance matrix attribute),
the specified quantile, and quantile-based cutoff, the equivalent \emph{p}-value, the cutoff based on
\code{alpha} (if specified), and the actual suggested cutoff. A message about the significance of
the quantile-based cutoff may be included.
}
\seealso{
\code{\link{gsnDistanceHistogram}}
}
