% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnDistanceHistogram.R
\name{gsnDistanceHistogram}
\alias{gsnDistanceHistogram}
\title{gsnDistanceHistogram}
\usage{
gsnDistanceHistogram(
  object,
  distance = NULL,
  dist.matrix = c("raw", "pared", "edges"),
  stat = "percent",
  colors = NULL,
  bins = 100,
  use_facets = FALSE,
  scales = "free",
  ncol = 1
)
}
\arguments{
\item{object}{A GSNData object}

\item{distance}{A distance, or even a character vector of distances, e.g. c("lf", "jaccard", "stlf").}

\item{dist.matrix}{The names of distance matrices, which can be "raw" for the distance stored in "matrix",
"pared" for the distances stored in "pared", and "edges" for the distances stored in "edges". In general,
the distances in "edges" will be the same as those in "pared", but this may not always be true.}

\item{stat}{Can be "percent", "count", "density", or "cumulative". This determines how the data are visualized.}

\item{colors}{Currently, this does nothing, but will eventually allow the user to specify custom colors. Stay
tuned.}

\item{bins}{The number of bins, for histograms ("percent" or "count").}

\item{use_facets}{A logical value that tells the function whether to generate a stacked or side-by-side plot
with multiple facets instead of plotting all distances on a single plot with the same scale. (default FALSE)}

\item{scales}{Specifies whether subplot scales should be "fixed", meaning that they share the same axis
scaling, "free_x", meaning that subplots have their own x coordinates, "free_y", meaning that subplots have
their own y coordinates, or "free" indicating that both axes are free. For more details, see documentation
for \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. This is only used when
\code{use_facets = TRUE} is specified. (Default "free")}

\item{ncol}{Specifies the number of columns of subplots. For more details, see documentation
for \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. This is only used when
\code{use_facets = TRUE} is specified. (Default 1)}
}
\value{
A ggplot2 graphical object.
}
\description{
Generate a Histogram of Distances
}
\details{
This function is useful for such purposes as assessing the effects of paring on the distribution of
distances.
}
\seealso{
\code{\link{plot.GSNData}}
\code{\link{gsnPlotNetwork}}
\code{\link{gsnHierarchicalDendrogram}}
}
