% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnImportFGSEA.R
\name{gsnImportFGSEA}
\alias{gsnImportFGSEA}
\title{gsnImportFGSEA}
\usage{
gsnImportFGSEA(
  object,
  pathways_data = NULL,
  filename = NULL,
  id_col = NULL,
  stat_col = NULL,
  sig_order = NULL,
  n_col = NULL,
  sep = "\\t"
)
}
\arguments{
\item{object}{A GSNData object.}

\item{pathways_data}{An (optional) data.frame containing the results of FGSEA analysis. (Either this or the
\code{filename} argument must be set.}

\item{filename}{An (optional) filename for data sets read from a text file containing FGSEA results. This is ignored
if the \code{pathways_data} argument is set.}

\item{id_col}{(optional) A character vector of length 1 indicating the name of the column used as a key for gene
sets or modules. This is normally the \code{pathway} field of FGSEA data, which must be the same as the names of gene
sets specified in the tmod object or in the list of gene set vectors specified with the \code{geneSetCollection}
argument used when building the gene set network. By default this value is \code{'pathway'}. The IDs must correspond
to the names of the gene sets provided, or an error will be thrown. \strong{NOTE:} In the tmod::tmodImportMSigDB function
provided by the tmod package, the default ID is an MSigDB accession, but FGSEA data sets do not use this accession.
The \code{NAME} column used in FGSEA results set corresponds instead to the \code{STANDARD_NAME} field in the MSigDB
XML database file. This \code{STANDARD_NAME} field is not preserved by the standard \code{tmod::tmodImportMSigDB}
utility function, but instead reformatted converting underscores to spaces and non-initial letters to lower case.
Therefore, when using FGSEA data sets with an MSigDB gene set collection imported using \code{tmod::tmodImportMSigDB}
the \code{NAME} fields need to be mapped to the \code{ID} or vice versa.}

\item{stat_col}{(optional) A character vector of length 1 indicating the name of the column used as a statistic
to evaluate the quality of pathways results. The function scans through possible \code{stat_col} values
(\code{"pval"}, and \code{"padj"}, and uses the first one it finds.}

\item{sig_order}{(optional) Either \code{'loToHi'} (default) or \code{'hiToLo'} depending on the statistic used to
evaluate pathways results.}

\item{n_col}{(optional) Specifies the column containing the number of genes in the gene set. Generally, this is the number
of genes in the gene set that are attested in an expression data set (Defaults to 'size').}

\item{sep}{A separator for text file import, defaults to "\\t". Ignored if \code{filename} is not specified.}
}
\value{
This returns a GSNData object containing imported pathways data.
}
\description{
Add FGSEA search data to a GSNData object, as generated by the the FGSEA package. The data set can be
either in the form of a data.frame or specified as import from a delimited text file. (See Details below)
}
\details{
Note: An error is thrown if all gene set IDs in the genePresenceAbsense are not present in the FGSEA \code{pathway} column.
However, if there are gene set IDs present in the pathways data that are absent from the \code{$genePresenceAbsence}
matrix, then this method emits a warning. It also checks for the standard FGSEA data set column names, and if some are
missing, it will emit a warning.
}
\seealso{
\code{\link{gsnAddPathwaysData}}
\code{\link{gsnImportCERNO}}
\code{\link{gsnImportGSEA}}
\code{\link{gsnImportGSNORA}}
\code{\link{gsnImportGenericPathways}}
}
