% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnPathways.R
\name{gsnPathways}
\alias{gsnPathways}
\alias{gsnPathways<-}
\title{gsnPathways}
\usage{
gsnPathways(object, ...)

gsnPathways(object, ...) <- value
}
\arguments{
\item{object}{A \code{GSNData} object with associated pathways data.}

\item{...}{The (optional) names of specific columns. If no column names are given, the whole
pathways data.frame is returned. When used as an LVALUE for assignment, column names must be
provided.}

\item{value}{When used in assignment, a list, data.frame, or vector RVALUE for assignment. If
value is a list or data.frame, the it must have as many members or columns as there are
named columns (see the ... argument above). If value is a vector, it is embedded into a list
and treated as a single column.}
}
\value{
\code{gsnPathways()} returns an imported pathways data.frame. When used as an LVALUE,
\code{`gsnPathways<-`()} returns a copied GSNData object to which RVALUES have been assigned.
}
\description{
This function (or more specifically, pair of functions) can be used to return an
imported pathways data.frame, or optionally, an individual column or multiple columns specified
by name. When used as the LVALUE of an assignment, a new column or columns can be assigned to
the pathways data.frame.
}
