% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{intV2Color}
\alias{intV2Color}
\title{intV2Color}
\usage{
intV2Color(rgb_v)
}
\arguments{
\item{rgb_v}{An integer or numeric vector of length 3 containing RGB channel
intensities from 0 to 255.}
}
\value{
A 24-bit color specification in the form "#FFFFFF".
}
\description{
Converts a numeric or integer vector of length 3 containing
RGB values in the range of 0 to 255 to 24 bit color specifications in the
form "#FFFFFF".
}
\examples{

col_v <- c( 255, 100, 240)
col <- intV2Color( col_v )

}
\seealso{
\code{\link[=color2IntV]{color2IntV()}}
}
