% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{logGraduate}
\alias{logGraduate}
\title{logGraduate}
\usage{
logGraduate(usr, .paradigm = c(1, 2, 5))
}
\arguments{
\item{usr}{A range of values to be represented by the gradations.}

\item{.paradigm}{A single numeric vector pattern to be used to generate the gradations. (Default is
c(1,2,5))}
}
\value{
A vector of gradations.
}
\description{
Based on a range of user supplied values (usr) and a (single) paradigm or pattern,
for the gradations, e.g. \code{c(1,2,5)}, \code{logGraduate()} returns a set of graduations.

This function does the main work of \code{\link{logGraduate}()}.
}
\seealso{
\code{\link[=logGraduations]{logGraduations()}}
}
