% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{logGraduations}
\alias{logGraduations}
\title{logGraduations}
\usage{
logGraduations(
  usr,
  nint = 5,
  log_margin = 10^0.1,
  .paradigms = list(c(1, 2, 5), c(1, 3), 1, 2, 3, 4, 5, 6, 8)
)
}
\arguments{
\item{usr}{A range of values to be represented by the gradations.}

\item{nint}{The approximate number of gradations to return (default 5).}

\item{log_margin}{This is a factor that extends the range of the gradation values returned.
All values are within the range of \eqn{min(usr)*log_margin -> max(usr)/log_margin}.
(Default is 10^{0.1} or about 25\%.)}

\item{.paradigms}{A list of numeric vector patterns to be used to generate the gradations. (Default is
list(c(1,2,5), c(1,3), 1, 2, 3, 4, 5, 6, 8 ))}
}
\value{
A vector of gradations.
}
\description{
This function is meant to do what axisTicks should do for data to be represented on a
log scale but performs poorly at. For numbers within the expected range of sizes of gene sets,
where the sizes are greater than 1, but vary by less than 3-4 orders of magnitude, this returns
an optimized set of gradations for a legend, based on a range of values provided by the user, the
total desired number of gradations, a log-scale margin.
}
\seealso{
\code{\link[=logGraduate]{logGraduate()}}
}
