% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnSubnetSummary.R
\name{lse}
\alias{lse}
\title{lse}
\usage{
lse(a, b)
}
\arguments{
\item{a}{A numeric log value.}

\item{b}{Another numeric log value.}
}
\value{
The log of the sum of the exponents of a and b.
}
\description{
Implements the "Log-Sum-Exponential trick" for calculating the log of the sums of
exponents without arithmetic underflows. This allows very small or very large numbers
to be summed in log space.
}
\details{
The "Log-Sum-Exponential trick" allows calculation of the log of the sum of
exponents of two arguments, such that exponentiation calls that would lead to
numeric underruns or overruns can be avoided, and it works as follows.

\deqn{\ln( e^{a} + e^{b} ) = a + \ln(1 + e^{b - a}) }

If \eqn{a} and \eqn{b} are sufficiently different, such that \eqn{b - a > 709}, then
R evaluates \eqn{exp(b - a)} as \eqn{Inf}. Hence \eqn{a + log( 1 + exp( b - a ) )}
is evaluated as \eqn{Inf}, which is incorrect. However, since

\deqn{\lim_{b - a\to\infty} a + \ln(1 + e^{b - a}) = b}

The function approximates the result in such cases as \eqn{b}.
}
\examples{

A <- 1E-40
B <- 3E-41

log_AB <- lse( log(A), log(B) )

# exp( log_AB ) == A + B

}
