% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat2UnitNormRank.R
\name{negative}
\alias{negative}
\title{negative}
\usage{
negative(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
The negative of the x argument, equal to \eqn{- x}.
}
\description{
This function returns the negative of a numeric or integer vector or matrix. This may be suitable as the
\code{matrix_scaling_fun()} argument for \code{gsnPareNetGenericHierarchic()} when being used with such distance metrics
as the Jaccard Index or Szymkiewicz–Simpson Overlap Coefficients to transform them into something more approximating a
distance in behavior.
}
\details{
This function also sets matrix or vector attributes appropriately for negation of the input.
}
