% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSNData.R
\name{print.GSNData}
\alias{print.GSNData}
\title{print.GSNData}
\usage{
\method{print}{GSNData}(x, ...)
}
\arguments{
\item{x}{A GSNData object.}

\item{...}{Additional parameters currently ignored, but included for consistency with generic print.}
}
\value{
Invisibly returns the GSNData object.
}
\description{
Print a short description of a \code{GSNData} object.
}
