% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{read_gmt}
\alias{read_gmt}
\title{read_gmt}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{The path to GMT file to parse.}
}
\value{
This returns a GSC (gene set collection) as a name list of vectors, where the names correspond to gene set
identifiers and the vectors are gene symbols.
}
\description{
This function parses a GMT file, documented
\href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:\emph{Gene_Matrix_Transposed_file_format}.28.2A.gmt.29}{here}.
}
\seealso{
\code{\link[=gsc2tmod]{gsc2tmod()}}
}
