% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termSummary.R
\name{score_word_clusters}
\alias{score_word_clusters}
\title{score_word_clusters}
\usage{
score_word_clusters(
  term_cluster,
  single_word_score = 1,
  additional_word_score = 1,
  priority_factor = 1.2,
  zero_score_words = get_zero_score_words()
)
}
\arguments{
\item{term_cluster}{A vector of terms corresponding to a group of descriptive gene signature names.
These names may contain underscores or spaces and be in either upper or lower case.}

\item{single_word_score}{The score awarded to a word cluster for a single non-zero score word. (default 1)}

\item{additional_word_score}{The score awarded to a word cluster for subsequent non-zero score words. (default 1)}

\item{priority_factor}{The factor by which successive occurrences of word clusters are attenuated. (default 1.2)}

\item{zero_score_words}{A list of zero-score words. (defaults to the value returned by
\code{get_zero_score_words()})}
}
\value{
A named numeric vector for which the values are scores for each word-cluster, and the names
are word clusters summarizing the input cluster of terms. The results are sorted by descending score,
such that the first name:value pair corresponds to the best/highest scoring word cluster and its
score.
}
\description{
This function takes a vector of terms (a 'term cluster') corresponding to a group of descriptive
gene signature names, and breaks them down into subgroupings "word clusters", which are then scored
on the basis of the number of times they appear, number of words comprising them, and by their
priority, with words earlier in the list generally receiving higher priority
(if \code{priority_factor > 1}). The score for single words can be set independently of additional
words (\code{single_word_score} and \code{additional_word_score}), but by default they are both 1.
Words for which no score is given (e.g. "gobp","of") can also be specified as a character vector.
This function is the basis of \code{\link{termSummary}()}.
}
\examples{
library( GSNA )
terms <- c("complement activation", "complement cascade", "complement binding",
            "wp complement activation classical pathway",
            "complement activation classical pathway",
            "humoral immune response mediated by circulating immunoglobulin",
            "comp pathway", "regulation of complement activation",
            "classic pathway", "complement activation alternative pathway",
            "initial triggering of complement",  "lectin pathway",
            "opsonin binding")
 word_cluster_scores <- score_word_clusters( terms )
 # Best summary word cluster:
 word_cluster_scores[1]
 # complement activation
 # 5.06769

 # With different scoring weights:
 score_word_clusters( terms,
                      single_word_score = 0.4,
                      additional_word_score = 1,
                      priority_factor = 1.1  )[1]

 # complement activation classical pathway
 # 4.876716


}
\seealso{
\code{\link{termSummary}}
}
