% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreKappaMatrix_C}
\alias{scoreKappaMatrix_C}
\title{scoreKappaMatrix_C}
\usage{
scoreKappaMatrix_C( geneSetCollection_m )
}
\arguments{
\item{geneSetCollection_m}{(required) A logical presence/absence matrix representation of a gene set collection
in which columns correspond to gene sets, rows correspond to genes and values are \code{TRUE} if a gene is present
in a gene set and \code{FALSE} otherwise. Row and column names correspond to gene symbols and gene set
identifiers, respectively. NOTE: for a typical GSNA analysis, this matrix would include only observed filtered
genes and significant gene set hits from pathways analysis. Using a matrix version of the full MSigDB without filtering
genes, for example, would likely be unworkably slow and memory intensive.}
}
\value{
A numerical matrix containing the specified kappa values for all non-self pairs.

The \code{distance} attribute in the output matrix is set to \code{'kappa'}.
}
\description{
Takes a presence/absence matrix with genes as the rows and modules as columns and calculates
a matrix of Cohen's kappa values.
}
\details{
Cohen's kappa statistic (\emph{κ}) is generally used to evaluate the concordance of predictors, with 1 being
perfect concordance, -1 being perfect negative concordance, and 0 indicating no significant
concordance.

\deqn{\biggl[\begin{matrix}a & b \\ c & d\end{matrix}\biggr]}

Given a background of observable genes and two gene sets, \emph{i} and \emph{j} that may overlap, this contingency
table is used to represent four numbers:

\itemize{
\item{\emph{a}: the number of genes observed in the background but not in \emph{i} or \emph{j}}
\item{\emph{b}: the number of observed genes in \emph{i} but not \emph{j}}
\item{\emph{c}: the number of observed genes in \emph{j} but not \emph{i} and}
\item{\emph{d}: the number of observed genes in both \emph{j} and \emph{i}, i.e. the overlap.}
}

Kappa is calculated as follows:

\deqn{
 \kappa = \frac{ 2 \cdot ( ( a \cdot d) - (b \cdot c) ) }{(a+b) \cdot (b+d) + (a+c) \cdot (c+d)}
}
}
\examples{

library( GSNA )

# Get the background of observable genes set from
# expression data:
gene_background <- toupper(rownames( Bai_empty_expr_mat ))

# Using the sample gene set collection **Bai_gsc.tmod**,
# generate a gene presence-absence matrix filtered for the
# ref.background of observable genes:
presence_absence.mat <-
  makeFilteredGenePresenceAbsenceMatrix( ref.background = gene_background,
                                         geneSetCollection = Bai_gsc.tmod )

kappa.mat <- scoreKappaMatrix_C( presence_absence.mat )

}
\seealso{
\code{\link{scoreJaccardMatrix_C}}
}
