% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termSummary.R
\name{termSummary}
\alias{termSummary}
\title{termSummary}
\usage{
termSummary(
  terms,
  group = NULL,
  sep = ",",
  convert_underscores = TRUE,
  single_word_score = 1,
  additional_word_score = 1,
  priority_factor = 1.2,
  .stopwords = get_stopwords(),
  .zero_score_words = get_zero_score_words(),
  num_combine = 1,
  make_unique = TRUE
)
}
\arguments{
\item{terms}{A character vector containing terms, specifically the names of
gene sets. If a \code{group} vector is supplied, then it is used to group terms
into term clusters. If no \code{group} vector is supplied, the function treats
each element of the character vector as a group, and splits on internal commas
(or an arbitrary character specified using the argument \code{sep}) for each
term cluster.}

\item{group}{An optional vector for grouping terms into term clusters.}

\item{sep}{Character to dilineate terms in a term cluster. (default ",")}

\item{convert_underscores}{Logical value telling the function whether or not to
convert underscores to spaces. (default TRUE)}

\item{single_word_score}{The numeric base score awarded for the first word of a
term cluster (default 1).}

\item{additional_word_score}{The numeric base score awarded for subsequent words
of a term cluster (default 1).}

\item{priority_factor}{Numeric value determining the the score advantage due
to priority. Successive occurrences of a term summary in a vector of terms
have their scores divided by this number, so terms later in the list are
increasingly penalized (default 1.2)}

\item{.stopwords}{A vector of stopwords to filter out of terms (defaults to the
output value of the private function get_stopwords().)}

\item{.zero_score_words}{A vector of score free words, words that do not
contribute to term scores. (defaults to the output value of the private
function get_zero_score_words().)}

\item{num_combine}{The number of (best) term summaries to combine for each term
cluster. (default 1).}

\item{make_unique}{Logical value that tells the function to make output term
summaries unique. (default: TRUE)}
}
\value{
A vector of summarized terms for each group.
}
\description{
This is a utility function that takes a character vector of terms
(generally the names of gene signatures) and optionally a grouping vector and
returns a vector of best guesses for the summary terms best matching the input
terms. The function works by applying a scoring algorithm that weights terms
according to the number of words, priority in the supplied list of terms, and
the number of times a word cluster is found in the list.
}
\examples{

library( 'GSNA' )
.terms = c( "GOBP_REGULATION_OF_NATURAL_KILLER_CELL_ACTIVATION", # 1...
            "GOBP_POSITIVE_REGULATION_OF_NATURAL_KILLER_CELL_ACTIVATION",
            "GOBP_NATURAL_KILLER_CELL_ACTIVATION",
            "GOBP_POSITIVE_REGULATION_OF_NATURAL_KILLER_CELL_PROLIFERATION",
            "GOBP_ANTIBODY_DEPENDENT_CELLULAR_CYTOTOXICITY",
            "GOBP_TYPE_II_HYPERSENSITIVITY",
            "GOBP_NATURAL_KILLER_CELL_PROLIFERATION",
            "GOBP_MYELOID_LEUKOCYTE_ACTIVATION",                 # 2...
            "GOBP_MYELOID_LEUKOCYTE_MEDIATED_IMMUNITY",
            "GOBP_NATURAL_KILLER_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",
            "GOBP_MYELOID_CELL_ACTIVATION_INVOLVED_IN_IMMUNE_RESPONSE",
            "GOBP_LEUKOCYTE_DEGRANULATION",
            "GOBP_NATURAL_KILLER_CELL_DEGRANULATION",
            "GOBP_PHAGOCYTOSIS",
            "GOBP_REGULATED_EXOCYTOSIS" )
.groups = c(1,1,1,1,1,1,1,2,2,2,2,2,2,2,2)
termSummary( terms = .terms, group = .groups )

# 1                                     2
# "natural killer cell" "cell activation involved immune response"

}
