% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{tmod2gsc}
\alias{tmod2gsc}
\title{tmod2gsc}
\usage{
tmod2gsc(tmod)
}
\arguments{
\item{tmod}{: a tmod or tmodGS object.}
}
\value{
The function returns a gene set collection as a named list of character vectors containing gene names.
}
\description{
Function takes a tmod or tmodGS object and converts it to a gene set collection. In the case of a
tmod object, the function merely extracts the \code{$MODULES2GENES} list of character vectors. In the case of
tmodGS objects, the list of vectors of numeric gene identifiers in \code{$gs2gv} is converted to a named list
of character vectors of gene names.
}
\seealso{
\code{\link[=gsc2tmod]{gsc2tmod()}}
}
