% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnSubnetsDotPlot.R
\name{gsnSubnetsDotPlot}
\alias{gsnSubnetsDotPlot}
\title{gsnSubnetsDotPlot}
\usage{
gsnSubnetsDotPlot(
  object,
  pathways.data = NULL,
  distance = NULL,
  id_col = NULL,
  stat_col = NULL,
  sig_order = NULL,
  n_col = NULL,
  subnet_col = "subnet",
  title_col = "ID",
  color_col = NULL,
  color_sig_order = NULL,
  group_by = "subnet",
  summarize = FALSE,
  preappend_subnet_in_summary = TRUE,
  summary_annotation_name = "Summary Annotation",
  x_transform = "identity",
  axis_font_size = 8,
  axis_label_font_size = NULL,
  x_axis_font_size = NULL,
  y_axis_font_size = NULL,
  x_axis_label_font_size = NULL,
  y_axis_label_font_size = NULL,
  colors = NULL,
  midpoint = NULL,
  label_width_chars = NULL,
  col_widths = c(neg_inf = 1, fin = 10, pos_inf = 1),
  interactive = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{object}{A GSNData object.}

\item{pathways.data}{A pathways data.frame. If not specified, this defaults to the pathways data.frame that
was imported prior to subnet assignment.}

\item{distance}{A specifier for distance metric. If not specified, this defaults to the return value of
\code{gsn_default_distance(object)}.}

\item{id_col}{Specifies the id column of the pathways data.frame. (default: value of
\code{object$pathways$id_col})}

\item{stat_col}{Specifies the id column of the pathways data.frame to be used for plotting the x axis.
(This is only used in very particular circumstances, and can generally be ignored.
default: value of \code{object$pathways$id_col})}

\item{sig_order}{Specifies the sig_order of the stat_col column of the pathways data.frame to be used for
plotting the x axis. (default: value of \code{object$pathways$sig_order})}

\item{n_col}{Specifies the name of the pathways data.frame column to be used for plotting the dot sizes.
Generally, this corresponds to gene set sizes, but other fields may be used for plotting.}

\item{subnet_col}{This is the name of the column to be used to group subnets. (default \code{'subnet'})}

\item{title_col}{This is the name of the column to be used to identify gene sets. (default \code{'ID'})}

\item{color_col}{This is the name of the column to be used for the color aesthetic. (default NULL)}

\item{color_sig_order}{This is the name of the column to be used to indicate the significance order of
for the color aesthetic. (default \code{'hiToLo'})}

\item{group_by}{This is the name of the column to be used to label subnets, indicated on the y-axis.
(default \code{'subnet'})}

\item{summarize}{Logical value indicating whether the function should use the \code{termSummary} feature
to generate descriptive names for subnets. This is currently an experimental feature. (default \code{FALSE})}

\item{preappend_subnet_in_summary}{Logical value indicating whether the function should preappend subnet
numbers to the subnet descriptors used for the y-axis. (default \code{TRUE})}

\item{summary_annotation_name}{A name for the the summarized annotation column, if one is used.
(default: \code{'Summary Annotation'})}

\item{x_transform}{The transform used for the x_axis. This would normally be either \code{'identity'} or
\code{'log10'}, but can take other values. (See the \code{transform} argument of \strong{ggplot2's}
\code{\link[ggplot2]{scale_x_continuous}()} function.}

\item{axis_font_size}{Specifies the size of the axis tick fonts in points. If set as NULL, then the
function attempts to pick an appropriate size based on the number of characters in the longest y
label, and the number of y values. (default 8)}

\item{axis_label_font_size}{Specifies the size of the axis label fonts in points (default 2 x axis_font_size).}

\item{x_axis_font_size}{Specifies the size of the x axis tick fonts in points (default axis_font_size).}

\item{y_axis_font_size}{Specifies the size of the y axis tick fonts in points (default axis_font_size).}

\item{x_axis_label_font_size}{Specifies the size of the x axis label fonts in points (default axis_label_font_size).}

\item{y_axis_label_font_size}{Specifies the size of the y axis label fonts in points (default axis_label_font_size).}

\item{colors}{Specifies a palette of colors to use. If NULL, the function attempts to select a color palette based
on whether the values in \code{color_col} are positive (white to red), negative (red to white), or include both
positive and negative values (red to white to blue). If a 3 color palette is specified, the \code{midpoint}
argument can be used to specify a midpoint value for a color scale. (see \code{midpoint} below.)}

\item{midpoint}{A value to be used to specify the middle value of a color scale. If the \code{midpoint} argument
is specified, and the \code{colors} argument contains exactly 3 colors, then the second one is used for the
middle color. This is useful for specifying a zero-value (or mid value) color.}

\item{label_width_chars}{Numeric value indicating the length of a y-axis gene set label to wrap. If \code{NULL}.
no wrapping is performed. (default NULL)}

\item{col_widths}{This function uses either the \strong{patchwork} package or the \strong{plotly} \code{subplot()} function
to create a plot from multiple subplots, such that -Inf or Inf can be represented as broken x-axis values. The
\code{col_widths} argument specifies the relative sizes of these subplots.}

\item{interactive}{A single logical value specifying that either a ggplot object should be returned
(FALSE, non-interactive) or alternatively an interactive \strong{plotly} object should be returned (TRUE).
(default FALSE).}

\item{width}{This specifies the width of the plotting area in inches. This is mainly to enable automatic
calculation of the font size, and does not determine the size of images generated by this function.
(default, value of \code{grDevices::dev.size()[1]})}

\item{height}{This specifies the height of the plotting area in inches. This is mainly to enable automatic
calculation of the font size, and does not determine the size of images generated by this function.
(default, value of \code{grDevices::dev.size()[2]})}
}
\value{
This function returns either a \strong{ggplot2} \code{ggplot} object (default) or a \code{plotly/htmlwidget}
if the \code{interactive=TRUE} argument is specified.
}
\description{
This function generates a \strong{ggplot2} or (if the \code{interactive=TRUE} argument is specified)
plotly dotplot representation of gene set clusters/subnets. Gene sets are represented on the y-axis,
whereas the x-axis position, color, and size of dots represent user-chosen statistics for each gene set.

NOTE: Unlike the \code{gsnPlotNetwork()} and \code{genHierarchicalDendrogram()} methods,
\code{gsnSubnetsDotPlot()} does not currently have direct support for output to a file. For outputting to
a file, use the \strong{ggplot2} \code{ggsave()} function for \strong{ggplot} objects, or for \strong{plotly} objects,
use the \code{saveWidget()} function from the \strong{htmlwidget} package.
}
