% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsnSubnetSummary.R
\name{lse}
\alias{lse}
\title{lse}
\usage{
lse(a, b)
}
\arguments{
\item{a}{A numeric log value.}

\item{b}{Another numeric log value.}
}
\value{
The log of the sum of the exponents of a and b.
}
\description{
Implements the "Log-Sum-Exponential trick" for calculating the log of the sums of
exponents without arithmetic underflows. This allows very small numbers to be
summed in log space.
}
\examples{

A <- 1E-40
B <- 3E-41

log_AB <- lse( log(A), log(B) )

# exp( log_AB ) == A + B


}
